third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/combine-multiple-policies-expected.txt

CONSOLE ERROR: line 11: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 14: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'none'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

This test checks that we enforce all the supplied policies. This test passes if it doesn't alert fail and if the style doesn't apply.
